Move tests to qs exporter test. Fix lint and type errors.
